+++ tests/testthat/test-form_list.R
@@ -10,18 +10,19 @@ test_that("form_list works", {
   testthat::expect_equal(class(fl), c("tbl_df", "tbl", "data.frame"))
   cn <- c(
     "name", "fid", "version", "state", "submissions", "created_at",
-    "created_by_id", "created_by", "updated_at", "last_submission", "hash"
+    "created_by_id", "created_by", "updated_at", "published_at",
+    "last_submission", "hash"
   )
   testthat::expect_equal(names(fl), cn)
 
   # Testing #86 form_list should work with draft forms present.
   # The above call to form_list worked so we test here that
   # the test forms contain a draft form.
-  # Draft forms have a NA hash and version.
+  # Update: https://github.com/ropensci/ruODK/issues/119
+  # Draft forms, and any forms from older Central versions
+  # have published_at = NA.
   fl %>%
-    dplyr::filter(
-      is.na(version) & is.na(hash)
-    ) %>%
+    dplyr::filter(is.na(published_at)) %>%
     nrow() %>%
     testthat::expect_gt(0)
 })